requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via httpx
atproto==0.0.58
    # via -r requirements.in
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via atproto
cryptography==43.0.3
    # via atproto
dnspython==2.7.0
    # via atproto
feedgen==1.0.0
    # via -r requirements.in
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.27.2
    # via
    #   atproto
    #   notion-client
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
libipld==3.0.0
    # via atproto
lxml==5.3.0
    # via feedgen
notion-client==2.3.0
    # via -r requirements.in
pycparser==2.22
    # via cffi
pydantic==2.10.6
    # via atproto
pydantic-core==2.27.2
    # via pydantic
python-dateutil==2.9.0.post0
    # via feedgen
python-dotenv==1.0.1
    # via -r requirements.in
python-slugify==8.0.4
    # via -r requirements.in
pytz==2025.1
    # via -r requirements.in
requests==2.32.3
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
text-unidecode==1.3
    # via python-slugify
typing-extensions==4.12.2
    # via
    #   anyio
    #   atproto
    #   pydantic
    #   pydantic-core
urllib3==2.3.0
    # via requests
websockets==13.1
    # via atproto
